+++ ICD.Connect.Partitioning/Devices/IoPartitionDevice.cs
@@ -223,11 +223,16 @@ namespace ICD.Connect.Partitioning.Devices
 			if (portId == null)
 				return null;
 
-			IIoPort port = factory.GetPortById((int)portId) as IIoPort;
-			if (port == null)
-				Logger.AddEntry(eSeverity.Error, "No Serial Port with id {0}", portId);
+			try
+			{
+				return factory.GetPortById((int)portId) as IIoPort;
+			}
+			catch (KeyNotFoundException)
+			{
+				Log(eSeverity.Error, "No Serial Port with id {0}", portId);
+			}
 
-			return port;
+			return null;
 		}
 
 		#endregion